replaced event types in separated folder
