+++ src/types/events/EventHtmlTargetValueStringType.ts
